RA-20: 1. Созданы пакеты data, ui, model. Файлы распределены по пакетам. 2. В пакете recipe создал ViewModel, в нем создал data class с постфиксом State в названии.
