+++ app/src/main/java/com/example/myrecipebook/model/Ingredient.kt
@@ -1,4 +1,4 @@
-package com.example.myrecipebook
+package com.example.myrecipebook.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
