+++ app/src/main/java/com/example/myrecipebook/ui/categories/CategoriesListAdapter.kt
@@ -1,18 +1,19 @@
-package com.example.myrecipebook
+package com.example.myrecipebook.ui.categories
 
 import android.graphics.drawable.Drawable
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.example.myrecipebook.databinding.ItemRecipeBinding
+import com.example.myrecipebook.databinding.ItemCategoryBinding
+import com.example.myrecipebook.model.Category
 import java.io.IOException
 
-class RecipesListAdapter(private val dataSet: List<Recipe>) :
-    RecyclerView.Adapter<RecipesListAdapter.ViewHolder>() {
+class CategoriesListAdapter(private val dataSet: List<Category>) :
+    RecyclerView.Adapter<CategoriesListAdapter.ViewHolder>() {
 
     interface OnItemClickListener {
-        fun onItemClick(recipeId: Int)
+        fun onItemClick(categoryId: Int)
     }
 
     private var itemClickListener: OnItemClickListener? = null
@@ -21,29 +22,34 @@ class RecipesListAdapter(private val dataSet: List<Recipe>) :
         itemClickListener = listener
     }
 
-    inner class ViewHolder(private val binding: ItemRecipeBinding) :
+    inner class ViewHolder(private val binding: ItemCategoryBinding) :
         RecyclerView.ViewHolder(binding.root) {
 
-        fun bind(recipe: Recipe) {
-            binding.tvRecipeTitle.text = recipe.title
+        fun bind(category: Category) {
+            binding.tvCategoryTitle.text = category.title
+            binding.tvCategoryDescription.text = category.description
 
             try {
-                val inputStream = itemView.context.assets.open(recipe.imageUrl)
+                val inputStream = itemView.context.assets.open(category.imageUrl)
                 val drawable = Drawable.createFromStream(inputStream, null)
-                binding.ivRecipeImage.setImageDrawable(drawable)
+                binding.ivCategoryImage.setImageDrawable(drawable)
             } catch (e: IOException) {
-                Log.e("RecipesAdapter", "Error loading image: ${recipe.imageUrl}", e)
+                Log.e(
+                    "CategoriesAdapter",
+                    "Error loading image: ${category.imageUrl}",
+                    e
+                )
             }
 
             binding.root.setOnClickListener {
-                itemClickListener?.onItemClick(recipe.id)
+                itemClickListener?.onItemClick(category.id)
             }
         }
     }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         val inflater = LayoutInflater.from(parent.context)
-        val binding = ItemRecipeBinding.inflate(inflater, parent, false)
+        val binding = ItemCategoryBinding.inflate(inflater, parent, false)
         return ViewHolder(binding)
     }
