+++ app/src/main/java/com/example/myrecipebook/ui/categories/CategoriesListFragment.kt
@@ -1,4 +1,4 @@
-package com.example.myrecipebook
+package com.example.myrecipebook.ui.categories
 
 import android.os.Bundle
 import android.util.Log
@@ -8,6 +8,12 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.commit
 import androidx.fragment.app.replace
+import com.example.myrecipebook.ARG_CATEGORY_ID
+import com.example.myrecipebook.ARG_CATEGORY_IMAGE_URL
+import com.example.myrecipebook.ARG_CATEGORY_NAME
+import com.example.myrecipebook.R
+import com.example.myrecipebook.ui.recipes.recipelist.RecipesListFragment
+import com.example.myrecipebook.data.STUB
 import com.example.myrecipebook.databinding.FragmentListCategoriesBinding
 
 class CategoriesListFragment : Fragment() {
